Re-organize storybook so that source code is visible (make canvas, with show code by default)
Organize storybooks

Breakup a little more

Remove text searching examples
