+++ package.json
@@ -39,12 +39,12 @@
     "@oclif/dev-cli": "^1.25.1",
     "@oclif/test": "^1.2.7",
     "@rescripts/cli": "^0.0.14",
-    "@storybook/addon-actions": "^6.3.4",
-    "@storybook/addon-docs": "^6.3.4",
-    "@storybook/addon-links": "^6.3.4",
-    "@storybook/addons": "^6.3.4",
+    "@storybook/addon-actions": "^6.3.7",
+    "@storybook/addon-docs": "^6.3.7",
+    "@storybook/addon-links": "^6.3.7",
+    "@storybook/addons": "^6.3.7",
     "@storybook/preset-create-react-app": "^3.2.0",
-    "@storybook/react": "^6.3.4",
+    "@storybook/react": "^6.3.7",
     "@testing-library/jest-dom": "^5.5.0",
     "@testing-library/react": "^11.0.0",
     "@types/base64-js": "^1.2.5",
