+++ products/jbrowse-react-linear-genome-view/stories/DefaultSessions.stories.mdx
@@ -94,5 +94,5 @@ const defaultSession = {
 Here is an example LGV with a long reads alignment track loaded in the `defaultSession`:
 
 <Canvas>
-  <Story id="linear-view--linear-view-with-long-reads" />
+  <Story id="linear-view--with-long-reads" />
 </Canvas>
