+++ products/jbrowse-react-linear-genome-view/stories/MoreExamples.stories.mdx
@@ -1,0 +1,22 @@
+import { Meta, Story, Canvas } from '@storybook/addon-docs/blocks'
+
+<Meta title="More Examples" />
+
+## Example showing two linear genome views on the same page
+
+This page of course has more than two instances, but this is just a quick code
+snippet to demonstrate
+
+<Canvas withSource="open">
+  <Story id="linear-view--two-linear-genome-views" />
+</Canvas>
+
+## Demo showing that outside styles don't conflict with the component
+
+The react component takes measures to avoid being affected by styles outside of
+it on the page. In this case, a font and a text align, which without measures
+could affect the jbrowse-react-linear-genome-view component, is fine here.
+
+<Canvas withSource="open">
+  <Story id="linear-view--with-outside-styling" />
+</Canvas>
