+++ products/jbrowse-react-linear-genome-view/stories/NextstrainDemo.stories.mdx
@@ -1,0 +1,268 @@
+import { Meta, Story, Canvas } from '@storybook/addon-docs/blocks'
+
+<Meta title="Nextstrain SARS-CoV-2 demo" />
+
+## Nextstrain SARS-CoV-2 example
+
+We created this demo to help show how a instance can be used to browse the
+SARS-CoV-2 genome similar to Nextstrain (https://nextstrain.org). It uses a
+color config to color gene's similar to the Nextstrain website, and also
+displays "Entropy score" to show diversity
+
+<Canvas>
+  <Story id="linear-view--nextstrain-example" />
+</Canvas>
+
+### Source
+
+This config makes use of the "FromConfigAdapter" where we supply the feature
+data to the system through JSON rather than from flatfiles like GFF3 and
+BigWig. The FromConfigAdapter can be useful to display on-the-fly analyses from
+your code or to avoid having to write data to the disk e.g. generate a
+BigWig/GFF3. See
+products/jbrowse-react-linear-genome-view/public/nextstrain_covid.json for
+details, summarized below here too
+
+```
+{
+  "assembly": {
+    "name": "SARS-CoV-2",
+    "sequence": {
+      "type": "ReferenceSequenceTrack",
+      "trackId": "SARS-CoV-2-ReferenceSequenceTrack",
+      "adapter": {
+        "type": "FromConfigSequenceAdapter",
+        "features": [
+          {
+            "refName": "SARS-CoV-2",
+            "uniqueId": "SARS-CoV-2",
+            "start": 0,
+            "end": 29903,
+            "seq": "...actual sequence of the SARS-CoV-2 genome here..."
+          }
+        ]
+      }
+    }
+  },
+  "tracks": [
+    {
+      "type": "FeatureTrack",
+      "name": "Nextstrain annotations",
+      "trackId": "nextstrain-annotations",
+      "assemblyNames": ["SARS-CoV-2"],
+      "category": ["Annotation"],
+      "adapter": {
+        "type": "FromConfigAdapter",
+        "features": [
+          {
+            "refName": "SARS-CoV-2",
+            "name": "E",
+            "uniqueId": 4,
+            "start": 26244,
+            "end": 26472,
+            "fill": "#D9AD3D"
+          },
+          {
+            "refName": "SARS-CoV-2",
+            "name": "M",
+            "uniqueId": 5,
+            "start": 26522,
+            "end": 27191,
+            "fill": "#5097BA"
+          },
+          {
+            "refName": "SARS-CoV-2",
+            "name": "N",
+            "uniqueId": 10,
+            "start": 28273,
+            "end": 29533,
+            "fill": "#E67030"
+          },
+          {
+            "refName": "SARS-CoV-2",
+            "name": "ORF1a",
+            "uniqueId": 0,
+            "start": 265,
+            "end": 13468,
+            "fill": "#8EBC66"
+          },
+          {
+            "refName": "SARS-CoV-2",
+            "name": "ORF1b",
+            "uniqueId": 1,
+            "start": 13467,
+            "end": 21555,
+            "fill": "#E59637"
+          },
+          {
+            "refName": "SARS-CoV-2",
+            "name": "ORF3a",
+            "uniqueId": 3,
+            "start": 25392,
+            "end": 26220,
+            "fill": "#AABD52"
+          },
+          {
+            "refName": "SARS-CoV-2",
+            "name": "ORF6",
+            "uniqueId": 6,
+            "start": 27201,
+            "end": 27387,
+            "fill": "#DF4327"
+          },
+          {
+            "refName": "SARS-CoV-2",
+            "name": "ORF7a",
+            "uniqueId": 7,
+            "start": 27393,
+            "end": 27759,
+            "fill": "#C4B945"
+          },
+          {
+            "refName": "SARS-CoV-2",
+            "name": "ORF7b",
+            "uniqueId": 8,
+            "start": 27755,
+            "end": 27887,
+            "fill": "#75B681"
+          },
+          {
+            "refName": "SARS-CoV-2",
+            "name": "ORF8",
+            "uniqueId": 9,
+            "start": 27893,
+            "end": 28259,
+            "fill": "#60AA9E"
+          },
+          {
+            "refName": "SARS-CoV-2",
+            "name": "ORF9b",
+            "uniqueId": 11,
+            "start": 28283,
+            "end": 28577,
+            "fill": "#D9AD3D"
+          },
+          {
+            "refName": "SARS-CoV-2",
+            "name": "S",
+            "uniqueId": 2,
+            "start": 21562,
+            "end": 25384,
+            "fill": "#5097BA"
+          }
+        ]
+      },
+      "displays": [
+        {
+          "type": "LinearBasicDisplay",
+          "displayId": "nextstrain-color-display",
+          "renderer": {
+            "type": "SvgFeatureRenderer",
+            "color1": "jexl:get(feature,'fill') || 'black'"
+          }
+        }
+      ]
+    },
+    {
+      "type": "QuantitativeTrack",
+      "name": "Entropy score",
+      "trackId": "entropy-score",
+      "assemblyNames": ["SARS-CoV-2"],
+      "category": ["Annotation"],
+      "adapter": {
+        "type": "FromConfigAdapter",
+        "features": [
+          {
+            "refName": "SARS-CoV-2",
+            "score": 0.002,
+            "start": 26252,
+            "end": 26253,
+            "uniqueId": "26252"
+          },
+          /*more scores for the entropy track here */
+        ]
+      }
+    }
+  ],
+  "location": "SARS-CoV-2:1..29,903",
+  "defaultSession": {
+    "name": "My session",
+    "view": {
+      "id": "linearGenomeView",
+      "type": "LinearGenomeView",
+      "tracks": [
+        {
+          "type": "QuantitativeTrack",
+          "configuration": "entropy-score",
+          "displays": [
+            {
+              "type": "LinearWiggleDisplay",
+              "displayId": "entropy-score-LinearWiggleDisplay",
+              "renderers": {
+                "DensityRenderer": {
+                  "type": "DensityRenderer"
+                },
+                "XYPlotRenderer": {
+                  "type": "XYPlotRenderer"
+                },
+                "LinePlotRenderer": {
+                  "type": "LinePlotRenderer"
+                }
+              }
+            }
+          ]
+        },
+        {
+          "type": "FeatureTrack",
+          "configuration": "nextstrain-annotations",
+          "displays": [
+            {
+              "type": "LinearBasicDisplay",
+              "configuration": "nextstrain-color-display"
+            }
+          ]
+        },
+        {
+          "type": "ReferenceSequenceTrack",
+          "configuration": "SARS-CoV-2-ReferenceSequenceTrack",
+          "displays": [
+            {
+              "type": "LinearReferenceSequenceDisplay",
+              "configuration": "SARS-CoV-2-ReferenceSequenceTrack-LinearReferenceSequenceDisplay"
+            }
+          ]
+        }
+      ]
+    }
+  }
+}
+
+```
+
+And then we launch it like this
+
+```
+  const { assembly, tracks, defaultSession } = nextstrainConfig
+  const state = createViewState({
+    assembly,
+    tracks,
+    defaultSession,
+    location: 'SARS-CoV-2:1..29,903',
+    onChange: patch => {
+      // eslint-disable-next-line no-console
+      console.log('patch', patch)
+    },
+    configuration: {
+      theme: {
+        palette: {
+          primary: {
+            main: '#5da8a3',
+          },
+          secondary: {
+            main: '#333',
+          },
+        },
+      },
+    },
+  })
+```
