+++ products/jbrowse-react-linear-genome-view/stories/Theming.stories.mdx
@@ -1,0 +1,16 @@
+import { Meta, Story, Canvas } from '@storybook/addon-docs/blocks'
+
+<Meta title="Theming" />
+
+## Theming
+
+Users can customize various aspects by providing a theme. We have a concept of
+"primary","secondary",and "tertiary", and "quaternary" colors for the app which
+by default are "dark blue", "purple", and "yellow". This example changes those,
+as well as the colors used for the DNA bases.
+
+See https://jbrowse.org/jb2/docs/config_guide#configuring-the-theme for more info
+
+<Canvas withSource="open">
+  <Story id="linear-view--custom-theme" />
+</Canvas>
